refactor: centralize component data and cleanup
