+++ src/app/data/brands.ts
@@ -1,0 +1,7 @@
+import brands1_d from "/brands/Brands1_d.svg";
+import brands2_d from "/brands/Brands2_d.svg";
+import brands1_m from "/brands/Brands1_m.svg";
+import brands2_m from "/brands/Brands2_m.svg";
+
+export const BRANDS_DESKTOP = [brands1_d, brands2_d];
+export const BRANDS_MOBILE = [brands1_m, brands2_m];
