+++ src/app/data/button.ts
@@ -1,0 +1,5 @@
+export const BUTTON_SIZE_CLASS: Record<"sm" | "md" | "lg", string> = {
+  sm: "btn--sm",
+  md: "btn--md",
+  lg: "btn--lg",
+};
