+++ src/app/data/expertiseSlides.ts
@@ -1,0 +1,43 @@
+import { type Slide } from "../../components/ui/Slider/Slider.tsx";
+
+export const EXPERTISE_SLIDES: Slide[] = [
+  {
+    title: "Анализ существующей инфраструктуры и проектирование IT-решений",
+    text:
+      "Анализ существующей инфраструктуры и проектирование IT-решений (скрол)" +
+      "Полная инвентаризация и оценка существующей IT-среды" +
+      "Что включает аудит: инвентаризация и оценка оборудования — системы безопасности, видеонаблюдение, СКУД, UPS, " +
+      "резервное питание, системы резервирования и хранения данных; детальный учёт серверов, рабочих станций, " +
+      "сетевого оборудования и периферии \n" +
+      "Анализ сетевой инфраструктуры — оценка топологии, пропускной способности, безопасности и устойчивости сети" +
+      "Проверка серверов — тестирование производительности, отказоустойчивости, актуальности ОС и оборудования" +
+      "Аудит информационной безопасности — выявление уязвимостей, проверка антивирусной защиты и политик доступа" +
+      "Анализ программного обеспечения — проверка лицензий, чистоты, актуальности и соответствия бизнес-задачам" +
+      "Оценка ИТ-сервисов — диагностика работы почты, файловых хранилищ, ERP/CRM, систем резервного копирования",
+    image: "/slides/checklist.png",
+    primary: { label: "Связаться с нами", href: "/contacts" },
+    secondary: { label: "Скачать чек-лист", href: "/files/checklist.pdf" },
+  },
+  {
+    title: "Консалтинг",
+    text:
+      "Изучаем объект, определяем технические возможности, выявляем ограничения и " +
+      "разрабатываем оптимальную архитектуру для внедрения IT-решений.\n" +
+      "Оптимизация затрат — анализ текущих расходов и предложения по экономии\n" +
+      "Планирование миграций — безопасный переход на новые технологии и платформы\n" +
+      "Масштабирование инфраструктуры — подготовка к росту бизнеса\n" +
+      "Повышение безопасности — внедрение современных методов защиты",
+    image: "/slides/consulting.png",
+    primary: { label: "Заказать консалтинг", href: "/services/consulting" },
+  },
+  {
+    title: "Подход IT Source",
+    text:
+      "Стандартизированные методики — проверенные подходы к аудиту и консалтингу\n" +
+      "Детальная документация — полные отчёты с практическими рекомендациями\n" +
+      "Опытная команда — специалисты с практикой в проектах разного масштаба\n" +
+      "Применимые решения — не просто отчёт, а ценные рекомендации для внедрения",
+    image: "/slides/way.png",
+    primary: { label: "Связаться с нами", href: "/contacts" },
+  },
+];
