+++ src/app/data/features.tsx
@@ -1,0 +1,24 @@
+import { FaStar, FaCogs, FaWrench, FaPhoneAlt } from "react-icons/fa";
+
+export const FEATURES = [
+  {
+    icon: <FaStar className="text-3xl md:text-4xl text-[#2b2b2b]" />,
+    title: "10+ лет опыта",
+    subtitle: "От SMB до enterprise проектов",
+  },
+  {
+    icon: <FaCogs className="text-3xl md:text-4xl text-[#2b2b2b]" />,
+    title: "Полный цикл",
+    subtitle: "От анализа до внедрения и поддержки",
+  },
+  {
+    icon: <FaWrench className="text-3xl md:text-4xl text-[#2b2b2b]" />,
+    title: "Под ключ",
+    subtitle: "Все работы силами одной команды",
+  },
+  {
+    icon: <FaPhoneAlt className="text-3xl md:text-4xl text-[#2b2b2b]" />,
+    title: "Техподдержка 24/7",
+    subtitle: "Проактивный мониторинг и SLA",
+  },
+];
