+++ src/app/data/hostingSlides.ts
@@ -1,0 +1,50 @@
+import { type Slide } from "../../components/ui/Slider/Slider.tsx";
+
+export const HOSTING_SLIDES: Slide[] = [
+  {
+    title: "Выделенные серверы и VPS",
+    text:
+      "Выделенные серверы — полный контроль над оборудованием и производительностью\n" +
+      "Виртуальные серверы (VPS) — гибкая конфигурация и масштабирование\n" +
+      "Управляемые решения — берём на себя администрирование и мониторинг\n" +
+      "Разные конфигурации — от веб‑серверов до высоконагруженных БД\n" +
+      "1C RDP — качественный доступ с соблюдением лицензирования по RDP/root",
+    image: "/slides/servers.png",
+    primary: { label: "Подробнее", href: "/services/servers" },
+    secondary: { label: "Узнать больше", href: "/contacts" },
+    background: "#06498233",
+  },
+  {
+    title: "Хранилища и резервное копирование",
+    text:
+      "Высокоскоростные SSD‑хранилища — для требовательных приложений\n" +
+      "Объёмные HDD‑массивы — для архива и долгого хранения\n" +
+      "Автоматическое резервное копирование — настроенные сценарии бэкапа\n" +
+      "Географически распределённые копии — защита от форс‑мажоров",
+    image: "/slides/storage.png",
+    primary: { label: "Связаться с нами", href: "/contacts" },
+    background: "#06498233",
+  },
+  {
+    title: "Частные облака и виртуальные сети",
+    text:
+      "Private Cloud — собственное облако на базе наших мощностей\n" +
+      "Виртуальные частные сети — безопасные соединения между локациями\n" +
+      "Гибридные решения — интеграция с вашей локальной инфраструктурой\n" +
+      "API‑управление — автоматизация развертывания и управления ресурсами",
+    image: "/slides/clouds.png",
+    primary: { label: "Связаться с нами", href: "/contacts" },
+    background: "#06498233",
+  },
+  {
+    title: "Поддержка и SLA",
+    text:
+      "Uptime 99.9% — гарантированная доступность сервисов\n" +
+      "Поддержка 24/7/365 — круглосуточный мониторинг и техподдержка\n" +
+      "Проактивное обслуживание — предупреждение проблем до их возникновения\n" +
+      "Прозрачная отчётность — детальная аналитика использования ресурсов",
+    image: "/slides/support.png",
+    primary: { label: "Связаться с нами", href: "/contacts" },
+    background: "#06498233",
+  },
+];
