+++ src/app/data/projectPhases.ts
@@ -1,0 +1,22 @@
+export type Stage = {
+  number: number;
+  title: string;
+  subtitle: string;
+};
+
+export const DEFAULT_STAGES: Stage[] = [
+  { number: 1, title: "Анализ требований", subtitle: "изучение задач и особенностей объекта" },
+  { number: 2, title: "Проектирование", subtitle: "создание технической документации и схем" },
+  { number: 3, title: "Поставка оборудования", subtitle: "подбор совместимых и надежных решений" },
+  {
+    number: 4,
+    title: "Монтаж и настройка",
+    subtitle: "профессиональная установка силами сертифицированных специалистов",
+  },
+  { number: 5, title: "Тестирование", subtitle: "комплексная проверка работоспособности систем" },
+  {
+    number: 6,
+    title: "Обучение и поддержка",
+    subtitle: "передача проекта заказчику с полной документацией, поддержка и техобслуживание",
+  },
+];
