+++ src/app/data/readySolutions.ts
@@ -1,0 +1,50 @@
+export type Solution = {
+  id: string;
+  title: string;
+  image: string;
+  modalImage?: string;
+  contactHref?: string;
+  bullets: string[];
+};
+
+export const DEFAULT_SOLUTIONS: Solution[] = [
+  {
+    id: "industry",
+    title: "Для производственных предприятий",
+    image: "/sectors/industry.png",
+    modalImage: "/sectors/industry-modal.png",
+    contactHref: "/#contact",
+    bullets: [
+      "Пропускная система для работников",
+      "Контроль доступа в производственные зоны",
+      "Учёт рабочего времени",
+      "Интеграция с системами безопасности",
+    ],
+  },
+  {
+    id: "hotel",
+    title: "Для гостиниц",
+    image: "/sectors/hotel.png",
+    modalImage: "/sectors/hotel-modal.png",
+    contactHref: "/#contact",
+    bullets: [
+      "Датчики атмосферы объекта — температура, влажность, качество воздуха",
+      "Умное освещение — автоматическое управление в зависимости от присутствия",
+      "Системы климат‑контроля — оптимизация энергопотребления",
+      "Мониторинг оборудования — предиктивная аналитика и предупреждение поломок",
+    ],
+  },
+  {
+    id: "business",
+    title: "Для бизнес центров",
+    image: "/sectors/bizcenter.png",
+    modalImage: "/sectors/bizcenter-modal.png",
+    contactHref: "/#contact",
+    bullets: [
+      "Управление парковкой",
+      "Система бронирования переговорных",
+      "Контроль доступа арендаторов",
+      "Мониторинг инженерных систем",
+    ],
+  },
+];
