+++ src/app/data/technology.ts
@@ -1,0 +1,5 @@
+import brands3_d from "/brands/Brands3_d.svg";
+import brands3_m from "/brands/Brands3_m.svg";
+
+export const TECHNOLOGY_DESKTOP = [brands3_d];
+export const TECHNOLOGY_MOBILE = [brands3_m];
