+++ src/components/Home/Carousels/HostingSection.tsx
@@ -1,57 +1,5 @@
-import Slider, { type Slide } from "../../ui/Slider/Slider";
-
-const hostingSlides: Slide[] = [
-  {
-    title: "Выделенные серверы и VPS",
-    text:
-      "Выделенные серверы — полный контроль над оборудованием и производительностью\n" +
-      "Виртуальные серверы (VPS) — гибкая конфигурация и масштабирование\n" +
-      "Управляемые решения — берём на себя администрирование и мониторинг\n" +
-      "Разные конфигурации — от веб‑серверов до высоконагруженных БД\n" +
-      "1C RDP — качественный доступ с соблюдением лицензирования по RDP/root",
-    image: "/slides/servers.png",
-    primary: { label: "Подробнее", href: "/services/servers" },
-    secondary: { label: "Узнать больше", href: "/contacts" },
-    background:
-      "#06498233",
-  },
-  {
-    title: "Хранилища и резервное копирование",
-    text:
-      "Высокоскоростные SSD‑хранилища — для требовательных приложений\n" +
-      "Объёмные HDD‑массивы — для архива и долгого хранения\n" +
-      "Автоматическое резервное копирование — настроенные сценарии бэкапа\n" +
-      "Географически распределённые копии — защита от форс‑мажоров",
-    image: "/slides/storage.png",
-    primary: { label: "Связаться с нами", href: "/contacts" },
-    background:
-      "#06498233",
-  },
-  {
-    title: "Частные облака и виртуальные сети",
-    text:
-      "Private Cloud — собственное облако на базе наших мощностей\n" +
-      "Виртуальные частные сети — безопасные соединения между локациями\n" +
-      "Гибридные решения — интеграция с вашей локальной инфраструктурой\n" +
-      "API‑управление — автоматизация развертывания и управления ресурсами",
-    image: "/slides/clouds.png",
-    primary: { label: "Связаться с нами", href: "/contacts" },
-    background:
-      "#06498233",
-  },
-  {
-    title: "Поддержка и SLA",
-    text:
-      "Uptime 99.9% — гарантированная доступность сервисов\n" +
-      "Поддержка 24/7/365 — круглосуточный мониторинг и техподдержка\n" +
-      "Проактивное обслуживание — предупреждение проблем до их возникновения\n" +
-      "Прозрачная отчётность — детальная аналитика использования ресурсов",
-    image: "/slides/support.png",
-    primary: { label: "Связаться с нами", href: "/contacts" },
-    background:
-      "#06498233",
-  },
-];
+import Slider from "../../ui/Slider/Slider";
+import { HOSTING_SLIDES } from "../../../app/data/hostingSlides";
 
 export default function HostingSection() {
   return (
@@ -61,7 +9,7 @@ export default function HostingSection() {
       className="mt-14 md:mt-24"
       heading="Аренда серверов"
       subheading="Готовые серверные мощности в надёжном ЦОД: производительные серверы, хранилища и сетевые ресурсы с гарантией SLA и поддержкой 24/7."
-      slides={hostingSlides}
+      slides={HOSTING_SLIDES}
     />
   );
 }
