+++ src/components/Home/Carousels/SolutionsSection.tsx
@@ -1,62 +1,5 @@
-import Slider, { type Slide } from "../../ui/Slider/Slider";
-
-const slides: Slide[] = [
-  {
-    title: "Слаботочные системы",
-    text:
-      "Комплексные решения для безопасности и контроля\n" +
-      "Видеонаблюдение — IP-камеры, аналитика, облачное хранение\n" +
-      "Системы контроля доступа (СКУД) — управление доступом с биометрией, лифтовые контроллеры и учет времени\n" +
-      "Охранно-пожарная сигнализация (ОПС) — защита периметра и помещений\n" +
-      "Системы оповещения — звуковое и световое информирование\n" +
-      "Телефония — современные IP-АТС и интеграции",
-    image: "/slides/low-current.png",
-    secondary: { label: "Скачать чек-лист", href: "/files/checklist.pdf" },
-    background:
-      "linear-gradient(96.31deg, rgba(6, 73, 130, 0.38) 23.99%, rgba(1, 25, 42, 0.38) 69.11%)",
-  },
-  {
-    title: "Сетевые решения",
-    text:
-      "Надежная и масштабируемая сетевая инфраструктура\n" +
-      "Структурированные кабельные системы (СКС) — проектирование и монтаж по международным стандартам\n" +
-      "Беспроводные сети Wi-Fi — бесшовное покрытие любой площади\n" +
-      "Маршрутизация и коммутация — оптимальная топология для вашего бизнеса\n" +
-      "VPN и удалённый доступ — безопасное подключение филиалов и сотрудников\n" +
-      "Системы безопасности сети — межсетевые экраны, IPS/IDS, сегментация",
-    image: "/slides/network.png",
-    primary: { label: "Связаться с нами", href: "/contacts" },
-    background:
-      "linear-gradient(281.87deg, rgba(6, 73, 130, 0.38) 45.02%, rgba(1, 25, 42, 0.38) 85.34%)",
-  },
-  {
-    title: "Серверные решения",
-    text:
-      "Производительная и отказоустойчивая серверная инфраструктура\n" +
-      "Виртуализация — VMware, Hyper-V, Proxmox для оптимизации ресурсов\n" +
-      "Active Directory — централизованное управление пользователями и доступом\n" +
-      "Системы резервного копирования — надёжная защита критически важных данных\n" +
-      "Кластеризация — высокая доступность сервисов\n" +
-      "Гибридные облачные решения — интеграция с Azure, AWS, Yandex Cloud",
-    image: "/slides/server.png",
-    primary: { label: "Подробнее", href: "/services/servers" },
-    background:
-      "linear-gradient(281.87deg, rgba(6, 73, 130, 0.38) 45.02%, rgba(1, 25, 42, 0.38) 85.34%)",
-  },
-  {
-    title: "Смарт-офисы и переговорные",
-    text:
-      "Технологичные решения для современного офиса\n" +
-      "Конференц-системы — видеостены, плазменные панели, презентации, звук\n" +
-      "Умное освещение и климат — автоматизация и энергосбережение\n" +
-      "Бронирование переговорных — интеграция с календарями и системами доступа\n" +
-      "Мультимедиа решения — интерактивные доски, проекторы, звук",
-    image: "/slides/office.png",
-    primary: { label: "Связаться с нами", href: "/contacts" },
-    background:
-      "linear-gradient(281.87deg, rgba(6, 73, 130, 0.38) 45.02%, rgba(1, 25, 42, 0.38) 85.34%)",
-  },
-];
+import Slider from "../../ui/Slider/Slider";
+import { SOLUTIONS_SLIDES } from "../../../app/data/solutionsSlides";
 
 export default function SolutionsSection() {
   return (
@@ -66,7 +9,7 @@ export default function SolutionsSection() {
       className="mt-20 md:mt-52"
       heading="Проектирование и внедрение решений"
       subheading="Проектируем слаботочные, сетевые и серверные системы, создаём комплексные проекты для офисной инфраструктуры и внедряем их под ключ."
-      slides={slides}
+      slides={SOLUTIONS_SLIDES}
     />
   );
 }
