+++ src/components/Home/Carousels/SpecialSolution.tsx
@@ -1,67 +1,12 @@
-import Slider, { type Slide } from "../../ui/Slider/Slider";
-
-const slides: Slide[] = [
-  {
-    title: "Готовые интеграции",
-    text:
-      "Системы видеонаблюдения и аналитики, используемые для управления IP-камерами анализа видеопотоков и обеспечения безопасности. Основные функции: распознавание лиц, номеров автомобилей, детекция движения, интеграция с другими системами.",
-    image: "/slides/integrations.png",
-    primary: { label: "Подробнее", href: "/services/integrations" },
-  },
-  {
-    title: "Экосистема UKEY",
-    text:
-      `Комплексная платформа учета, контроля и управления доступом
-- Централизованное управление — единая система для всех объектов и локаций
-- Учет времени и доступа — контроль рабочего времени и перемещений сотрудников
-- Интеграция с HR-системами — синхронизация с кадровыми базами данных
-- API для разработчиков — возможность интеграции с любыми внешними системами
-- Мобильное приложение — управление и мониторинг с мобильных устройств`,
-    image: "/slides/ukey.png",
-  },
-  {
-    title: "R‑Keeper, Game‑Keeper",
-    text:
-      `- Система автоматизации для ресторанов и кафе. Включает модули для работы официантов, кассиров, менеджеров. Обеспечивает учет заказов, управление складом, финансовую отчетность и интеграцию с онлайн‑заказами (Click & Collect).
-- Решение для автоматизации игровых и развлекательных центров. Использует пластиковые карты с магнитной полосой вместо жетонов. Позволяет управлять игровыми автоматами, отслеживать использование оборудования в реальном времени, интегрируется с ресторанными системами (например, R‑Keeper) и поддерживает программы лояльности.`,
-    image: "/slides/rkeeper.png",
-    primary: { label: "Подробнее", href: "/services/rkeeper" },
-  },
-  {
-    title: "Shelter",
-    text:
-      "Shelter PMS — система управления отелем: прием бронирований с сайта и агрегаторов, автоматизация работы персонала, контроль всех бизнес‑процессов.",
-    image: "/slides/shelter.png",
-    primary: { label: "Подробнее", href: "/services/shelter" },
-  },
-  {
-    title: "Управление объектами и процессами",
-    text:
-      `Автоматизация ключевых бизнес‑процессов
-- Система бронирования — переговорные, конференц‑залы
-- Управление парковкой — контроль доступа, оплата, мониторинг занятости
-- Посетительский учет — регистрация гостей, выдача временных пропусков
-- Управление ресурсами — планирование использования помещений и оборудования`,
-    image: "/slides/management.png",
-  },
-  {
-    title: "IoT и умная автоматизация",
-    text:
-      `Интернет вещей для современного бизнеса
-- Датчики атмосферы объекта — температура, влажность, качество воздуха
-- Умное освещение — автоматическое управление в зависимости от присутствия
-- Системы климат‑контроля — оптимизация энергопотребления
-- Мониторинг оборудования — предиктивная аналитика и предупреждение поломок`,
-    image: "/slides/iot.png",
-  },
-];
+import Slider from "../../ui/Slider/Slider";
+import { SPECIAL_SOLUTION_SLIDES } from "../../../app/data/specialSolutionSlides";
 
 export default function SpecialSolutions() {
   return (
     <Slider
       heading="Специальные решения"
       subheading="Уникальные продукты и интеграции от IT Source. Собственные разработки и специализированные решения для автоматизации бизнес‑процессов, интеграции систем и создания умных пространств."
-      slides={slides}
+      slides={SPECIAL_SOLUTION_SLIDES}
     />
   );
 }
