+++ src/components/Home/ProjectPhases/ProjectPhases.tsx
@@ -1,51 +1,11 @@
 import { SectionTitle } from "../../ui";
-
-export type Stage = {
-  number: number;
-  title: string;
-  subtitle: string;
-};
+import { DEFAULT_STAGES, type Stage } from "../../../app/data/projectPhases";
 
 export type ProjectPhasesProps = {
   heading?: string;
   stages?: Stage[];
   className?: string;
 };
-
-const DEFAULT_STAGES: Stage[] = [
-  {
-    number: 1,
-    title: "Анализ требований",
-    subtitle: "изучение задач и особенностей объекта",
-  },
-  {
-    number: 2,
-    title: "Проектирование",
-    subtitle: "создание технической документации и схем",
-  },
-  {
-    number: 3,
-    title: "Поставка оборудования",
-    subtitle: "подбор совместимых и надежных решений",
-  },
-  {
-    number: 4,
-    title: "Монтаж и настройка",
-    subtitle:
-      "профессиональная установка силами сертифицированных специалистов",
-  },
-  {
-    number: 5,
-    title: "Тестирование",
-    subtitle: "комплексная проверка работоспособности систем",
-  },
-  {
-    number: 6,
-    title: "Обучение и поддержка",
-    subtitle:
-      "передача проекта заказчику с полной документацией, поддержка и техобслуживание",
-  },
-];
 
 export default function ProjectPhases({
   heading = "Этапы реализации проекта",
